Move the MiroTransformable code into a separate file
